vep_service/requirements.txt
# VEP Annotation Service Dependencies

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
aiosqlite>=0.19.0

# VEP integration
pyyaml>=6.0.1

# HTTP client
httpx>=0.25.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development
ruff>=0.1.6
mypy>=1.7.1